Bootstrapping and noise management stuff (#133)
* Clean up debugging globals between test runs - fixes segfault when calling ./bin/runTests

* Squashed commit of the following:

- Introduced HElib-specific exceptions

- Added HElib-specific assertion functions which throw the above

- Replaced uses of C's assert in favour of HElib assertions

- Replaced uses of NTL Error functions with throwing HElib exceptions

- Replaced std::exception usage with HElib exceptions

* Refactored the header files to have consistent header guards

* Renamed cloned_ptr.h to clonedPtr.h

* update README.md

moved HElib to under the github HomEnc organization

* Refactored deprecated gtest macros

* changed include to clonedPtr.h

* Remove duplicated example file
Renamed cloned_ptr.h in CMakeLists.txt

* new PGFFT module with partial integration into helib

* Added Victor new files to CMakeLists.txt

* Fixed old bootstrappin legacy/ported tests

* Fixed case typo

+++ src/binio.cpp
@@ -241,6 +241,6 @@ template<> void write_raw_vector<double>(ostream& str, const vector<double>& v)
   long sz = v.size();  
   write_raw_int(str, sz); 
 
-  for(long n: v)
+  for(double n: v)
     write_raw_double(str, n); 
 };
